# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Fair Play Chile",
    "version": "1.0.0"
  },
  "paths": {
...
    "/partidos/buscar": {
      "get": {
        "summary": "Busca partidos dinámicamente",
        "description": "Permite filtrar partidos por cualquier campo disponible en la tabla.",
        "tags": [
          "Partidos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtra por ID del partido"
          },
          {
            "in": "query",
            "name": "owner_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtra por ID del propietario"
          },
          {
            "in": "query",
            "name": "nombre",
            "schema": {
              "type": "string"
            },
            "description": "Filtra por nombre del partido"
          },
          {
            "in": "query",
            "name": "fecha",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filtra por fecha del partido"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de partidos que cumplen con los filtros",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "partidos": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al buscar partidos"
          }
        }
      }
    },
...
    "/partidos/solicitudes/pendientes": {
      "get": {
        "summary": "Listar solicitudes de ingreso pendientes",
        "description": "Retorna las solicitudes pendientes para los partidos de un dueño específico con paginación.",
        "tags": [
          "Partidos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "owner_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID del dueño de los partidos"
          },
          {
            "in": "query",
            "name": "next",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Token de paginación encriptado"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de solicitudes pendientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "solicitudes": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "total_solicitudes": {
                      "type": "integer",
                      "description": "Total de solicitudes pendientes encontradas"
                    },
                    "next": {
                      "type": "string",
                      "description": "Token de paginación para siguiente página"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Faltan parámetros"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      }
    },
...
  },
...
}